Bugs backend

+++ src/Controller/AppController.php
@@ -365,18 +365,21 @@ class AppController extends AbstractController
 
     function rrmdir($src)
     {
-        $dir = opendir($src);
-        while (false !== ($file = readdir($dir))) {
-            if (($file != '.') && ($file != '..')) {
-                $full = $src . '/' . $file;
-                if (is_dir($full)) {
-                    rrmdir($full);
-                } else {
-                    unlink($full);
-                }
-            }
+        if (file_exists ( $src ) ){
+            $dir = opendir($src);
+                    while (false !== ($file = readdir($dir))) {
+                        if (($file != '.') && ($file != '..')) {
+                            $full = $src . '/' . $file;
+                            if (is_dir($full)) {
+                                rrmdir($full);
+                            } else {
+                                unlink($full);
+                            }
+                        }
+                    }
+                    closedir($dir);
+                    rmdir($src);
         }
-        closedir($dir);
-        rmdir($src);
+       
     }
 }
