+++ src/Controller/UserController.php
@@ -29,7 +29,7 @@ class UserController extends AbstractController
         ]);
     }
   /**
-     * @Route("/panel", name="user_index", methods={"GET"})
+     * @Route("/panel", name="admin_panel", methods={"GET"})
      */
     public function panelAdmin()
     {
@@ -163,19 +163,20 @@ class UserController extends AbstractController
     }
 
     function rrmdir($src) {
-        $dir = opendir($src);
-        while(false !== ( $file = readdir($dir)) ) {
-            if (( $file != '.' ) && ( $file != '..' )) {
-                $full = $src . '/' . $file;
-                if ( is_dir($full) ) {
-                    rrmdir($full);
-                }
-                else {
-                    unlink($full);
-                }
-            }
+        if (file_exists ( $src ) ){
+            $dir = opendir($src);
+                    while (false !== ($file = readdir($dir))) {
+                        if (($file != '.') && ($file != '..')) {
+                            $full = $src . '/' . $file;
+                            if (is_dir($full)) {
+                                rrmdir($full);
+                            } else {
+                                unlink($full);
+                            }
+                        }
+                    }
+                    closedir($dir);
+                    rmdir($src);
         }
-        closedir($dir);
-        rmdir($src);
     }
 }
